requirements.txt
# Core AI/ML Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# HTTP & Async
httpx>=0.27.0
aiofiles>=24.1.0

# Database & Caching
redis[hiredis]>=5.1.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development Tools
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
black>=24.8.0
ruff>=0.7.0
mypy>=1.11.0
pre-commit>=3.8.0
